lambda/requirements.txt
python-telegram-bot~=20.0
requests
boto3
wget
tls_client
websockets